+++ src/apify_client/_resource_clients/key_value_store.py
@@ -906,8 +906,6 @@ class KeyValueStoreClientAsync(ResourceClientAsync):
         """
         metadata = await self.get(timeout=timeout)
 
-        keys_public_url = urlparse(self._build_url('keys'))
-
         request_params = self._build_params(
             limit=limit,
             exclusiveStartKey=exclusive_start_key,
@@ -924,6 +922,7 @@ class KeyValueStoreClientAsync(ResourceClientAsync):
             request_params['signature'] = signature
 
         keys_public_url = urlparse(self._build_url('keys', public=True))
+
         filtered_params = {k: v for k, v in request_params.items() if v is not None}
         if filtered_params:
             keys_public_url = keys_public_url._replace(query=urlencode(filtered_params))